+++ test/basepack_test_app/spec/active_record_adapter_spec.rb
@@ -8,6 +8,6 @@ describe Netzke::Basepack::DataAdapters::ActiveRecordAdapter do
 
   it "should build a list of model attributes" do
     adapter = Netzke::Basepack::DataAdapters::ActiveRecordAdapter.new(Issue)
-    adapter.model_attributes.map{|a| a[:name]}.should == %w[id title developer__name project__title status__id created_at updated_at]
+    adapter.model_attributes.map{|a| a[:name]}.should == %w[id title developer__name project__title status__id created_at updated_at].map(&:to_sym)
   end
 end